line color in session detail fragment fixed. Also made a bunch of small fixes such as checking for correct types, converting values to the desired format etc and fixed some UI stuff. Also cleaned up some code

+++ app/src/main/java/com/georgv/sporttrackerapp/customHandlers/TypeConverterUtil.kt
@@ -1,16 +1,7 @@
 package com.georgv.sporttrackerapp.customHandlers
 
 import androidx.room.TypeConverter
-import com.georgv.sporttrackerapp.data.LocationPoint
-import com.google.gson.Gson
-import kotlinx.serialization.decodeFromString
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
 import java.util.*
-
-import com.google.gson.reflect.TypeToken
-import org.osmdroid.util.GeoPoint
-import java.lang.reflect.Type
 
 
 class TypeConverterUtil {
@@ -24,16 +15,11 @@ class TypeConverterUtil {
         return date.time
     }
 
-
     fun msToKmhConverter(speed: Float): Float {
-
         return String.format(null,"%.2f", speed * 3.6f).toFloat()
     }
 
-    fun meterToKilometerConverter(distance: Double): Double {
+    fun meterToKilometerConverter(distance: Float): Double {
         return String.format(null,"%.2f", distance / 1000).toDouble()
     }
-
-
-
 }
